+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes to be released in next version
  * 
 
 🙌 Improvements
+ * Jitsi: Use Jitsi server from homeserver's Well Known, if present, to create conferences (#3158).
  * RoomMemberDetailsVC: Enable / disable "Hide all messages from this user" from settings (#4281).
  * SettingsVC: Show / hide NSFW and decrypted content options from build settings (#4290).